@@ -1,61 +1,60 @@
-import { initAudioSources } from './audio_sources'; //sounds';
+import { initAudioSources } from "./audio_sources"; //sounds';
 
 export let ctx;
 export let masterGain;
 export let _settings = {
-  osc3: ["sine","sine","sine"],
-  overtone: [0.8,0.3,0.1],
-  detune: [0,1,2],
-  delay: [0,1,1],
-  LPF: { frequency: 2000,Q: 3 },
-  HPF: { frequency: 60,Q: 3 },
+  osc3: ["sine", "sine", "sine"],
+  overtone: [0.8, 0.3, 0.1],
+  detune: [0, 1, 2],
+  delay: [0, 1, 1],
+  LPF: { frequency: 2000, Q: 3 },
+  HPF: { frequency: 60, Q: 3 },
 
-  eqHZs: [62.5,125,250,500,1000],
-  adsr: [0.02,0.2,0.5,0.2,1.4],
-  LFO1: { frequeycy: 60,target: null },
-  LFO2: { frequeycy: 60,target: null },
-  compression: { threshold: -50,ratio: 5,preAmp: 1,postAmp: 1 },
+  eqHZs: [62.5, 125, 250, 500, 1000],
+  adsr: [0.02, 0.2, 0.5, 0.2, 1.4],
+  LFO1: { frequeycy: 60, target: null },
+  LFO2: { frequeycy: 60, target: null },
+  compression: { threshold: -50, ratio: 5, preAmp: 1, postAmp: 1 },
 };
 export let passthrough;
 export let noteCache;
 
-let keyCounter = 0,fftTimer = null;
-let compressor,analyser,preAmp,postAmp;
+let keyCounter = 0,
+  fftTimer = null;
+let compressor, analyser, preAmp, postAmp;
 
 export let instruments = {};
 const activeNotes = {};
+const useEvent = new BroadcastChannel("userEvent");
 const wschannel = new BroadcastChannel("wschannel");
 wschannel.onmessage = handleWs;
 
-async function handleWs({ data }) {
-  if (data.cmd && data.cmd === "updateSetting") {
-    const { key,idx,value } = data;
+async function handleWs({ data: { cmd, key, idx, value } }) {
+  if (cmd === "updateSetting") {
     _settings[key][idx] = value;
     value = parseFloat(value);
-    if (key == 'compression') {
-      if (idx === 'threshold') {
-        compressor.threshold
-          .setValueAtTime(value,ctx.currentTime);
+    if (key == "compression") {
+      if (idx === "threshold") {
+        compressor.threshold.setValueAtTime(value, ctx.currentTime);
       }
-      if (idx === 'ratio') {
-        compressor.ratio.setValueAtTime(value,ctx.currentTime);
-
+      if (idx === "ratio") {
+        compressor.ratio.setValueAtTime(value, ctx.currentTime);
       }
-      if (idx === 'preAmp') {
-        preAmp.gain.setValueAtTime(value,ctx.currentTime);
-
+      if (idx === "preAmp") {
+        preAmp.gain.setValueAtTime(value, ctx.currentTime);
       }
-      if (idx === 'postAmp') {
-        postAmp.gain.setValueAtTime(value,ctx.currentTime);
-
+      if (idx === "postAmp") {
+        postAmp.gain.setValueAtTime(value, ctx.currentTime);
       }
     }
   }
+}
+async function handleUserEvent({ data }) {
   if (data.cmd && (data.cmd === "keyboard" || data.cmd === "playback")) {
     await getContext();
     data.instrument = data.instrument || "piano";
 
-    const { freq,index,time,type,instrument } = data;
+    const { freq, index, time, type, instrument } = data;
     switch (type) {
       case "keydown":
       case "keypress":
@@ -78,35 +77,35 @@ async function handleWs({ data }) {
         break;
     }
   }
-};
+}
 
 const fftc = new BroadcastChannel("fftc");
 
-
-
-
 export async function getContext() {
+  if (ctx && ctx.state !== "running") {
+    await ctx.resume();
+    return ctx;
+  }
   if (ctx && postAmp) return ctx;
 
   ctx = ctx || new AudioContext();
   if (ctx.state === "paused") ctx.resume();
 
-  masterGain = masterGain || new GainNode(ctx,{ gain: 1 });
-  const sounds = await initAudioSources(ctx,masterGain,_settings);
+  masterGain = masterGain || new GainNode(ctx, { gain: 1 });
+  const sounds = await initAudioSources(ctx, masterGain, _settings);
 
-  Object.keys(sounds).map(functionName => {
+  Object.keys(sounds).map((functionName) => {
     instruments[functionName] = sounds[functionName];
-  })
+  });
 
-
-  passthrough = await require("./load-processor").loadProcessor(ctx,"pass-through");
-  compressor = new DynamicsCompressorNode(ctx,{
+  passthrough = await require("./load-processor").loadProcessor(ctx, "pass-through");
+  compressor = new DynamicsCompressorNode(ctx, {
     threshold: _settings.compression.threshold,
     radio: 4,
   });
-  analyser = new AnalyserNode(ctx,{ fftSize: 256,smoothingTimeConstant: 0.1 });
-  preAmp = new GainNode(ctx,{ gain: 1 });
-  postAmp = new GainNode(ctx,{ gain: 1 });
+  analyser = new AnalyserNode(ctx, { fftSize: 256, smoothingTimeConstant: 0.1 });
+  preAmp = new GainNode(ctx, { gain: 1 });
+  postAmp = new GainNode(ctx, { gain: 1 });
   masterGain
     .connect(preAmp)
     .connect(compressor)
@@ -118,9 +117,8 @@ export async function getContext() {
   return ctx;
 }
 
-
 const fftLoop = () => {
-  const dataArray = new Uint8Array(analyser.fftSize);
+  const dataArray = new Float32Array(analyser.fftSize);
   analyser.getByteFrequencyData(dataArray);
   fftc.postMessage({
     time: ctx.currentTime,
